change sign in module to spring-social-github for selectable login time

+++ jandy-server/src/main/java/io/jandy/domain/UserRepository.java
@@ -12,5 +12,5 @@ import java.util.List;
  */
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
-  User findByGitHubId(int id);
+  User findByGitHubId(long id);
 }
